fix(components): Reduce execution time using Components

+++ executor/games/delete/delete_games_executor.py
@@ -3,7 +3,7 @@
 
 import os
 from executor.games.abstract_games_executor import AbstractGamesExecutor
-from libraries.constants.constants import Action, Constants
+from libraries.constants.constants import Action, Component, Constants
 from libraries.context.context import Context
 from libraries.file.file_helper import FileHelper
 
@@ -19,11 +19,25 @@ class DeleteGamesExecutor(AbstractGamesExecutor):
     def do_execution(self, item: dict):
         """Do execution for an item"""
 
-        # Install game
-        FileHelper.delete_folder(
-            folder_path=os.path.join(
-                Context.get_games_path(),
-                Context.get_selected_platform().value,
-                item[Constants.UI_TABLE_KEY_COL_ID]
-            )
+        # Retrieve game folder
+        game_folder = os.path.join(
+            Context.get_games_path(),
+            Context.get_selected_platform().value,
+            item[Constants.UI_TABLE_KEY_COL_ID]
         )
+
+        # Delete game if all components requested
+        if Component.ROM in Context.get_selected_components():
+            FileHelper.delete_folder(
+                folder_path=game_folder
+            )
+            return
+
+        # Delete media if requested
+        if Component.MEDIA in Context.get_selected_components():
+            FileHelper.delete_folder(
+                folder_path=os.path.join(
+                    game_folder,
+                    self.MEDIA_FOLDER_NAME
+                )
+            )
