+++ manager/batocera/batocera_manager.py
@@ -2,13 +2,17 @@
 """Manager for the Software BATOCERA"""
 
 import os
-from libraries.constants.constants import Constants, Media, Platform, Software
+from libraries.constants.constants import Component, Constants, Media, Platform, Software
+from libraries.context.context import Context
 from libraries.file.file_helper import FileHelper
 from libraries.xml.xml_helper import XmlHelper
 from manager.abstract_manager import AbstractManager
 
 # pylint: disable=too-many-arguments
 # pylint: disable=too-many-positional-arguments
+# pylint: disable=too-many-locals
+# pylint: disable=too-many-branches
+# pylint: disable=too-many-statements
 
 
 class BatoceraManager(AbstractManager):
@@ -232,32 +236,35 @@ class BatoceraManager(AbstractManager):
         """Uninstall game"""
 
         # Delete media files
-        media_files = self.retrieve_media_files(
-            platform=platform,
-            game_item=game_item
-        )
-        for media_file in media_files.values():
-            FileHelper.delete_file(
-                file_path=media_file
-            )
-
-        # Delete rom file
-        FileHelper.delete_file(
-            file_path=self.retrieve_rom_file(
+        if Component.MEDIA in Context.get_selected_components():
+            media_files = self.retrieve_media_files(
                 platform=platform,
                 game_item=game_item
             )
-        )
+            for media_file in media_files.values():
+                FileHelper.delete_file(
+                    file_path=media_file
+                )
 
-        # Delete tag for the game
-        return XmlHelper.delete_tag(
-            xml_file_path=self.__retrieve_game_list_xml_path(
-                platform=platform
-            ),
-            parent_tag=self.__TAG_GAMES,
-            tag=self.__TAG_GAME,
-            criteria=self.__build_game_criteria(game_item)
-        )
+        # Delete rom file
+        if Component.ROM in Context.get_selected_components():
+            FileHelper.delete_file(
+                file_path=self.retrieve_rom_file(
+                    platform=platform,
+                    game_item=game_item
+                )
+            )
+
+            XmlHelper.delete_tag(
+                xml_file_path=self.__retrieve_game_list_xml_path(
+                    platform=platform
+                ),
+                parent_tag=self.__TAG_GAMES,
+                tag=self.__TAG_GAME,
+                criteria=self.__build_game_criteria(game_item)
+            )
+
+        return True
 
     def install_game(
         self,
@@ -278,17 +285,20 @@ class BatoceraManager(AbstractManager):
         # Initialize fields to add
         fields_to_add = {}
 
-        # Copy the rom
+        # Retrieve rom file
         batocera_rom_file = os.path.join(
             self._folder_path,
             self.__PATH_ROMS,
             self.__PLATFORM_DICT_INV.get(platform, ''),
             FileHelper.retrieve_file_name(rom_file)
         )
-        FileHelper.copy_file(
-            source_file_path=rom_file,
-            destination_file_path=batocera_rom_file
-        )
+
+        # Copy the rom
+        if Component.ROM in Context.get_selected_components():
+            FileHelper.copy_file(
+                source_file_path=rom_file,
+                destination_file_path=batocera_rom_file
+            )
 
         # Add field for the rom
         fields_to_add[self.__TAG_PATH] = self.__FILE_PREFIX
@@ -325,7 +335,8 @@ class BatoceraManager(AbstractManager):
                 media_folder = 'videos'
 
             # Copy media file in software
-            file_name = FileHelper.retrieve_file_basename(batocera_rom_file)
+            file_name = FileHelper.retrieve_file_basename(
+                batocera_rom_file)
             file_name += '-'
             file_name += key
             file_name += FileHelper.retrieve_file_extension(media_file)
@@ -336,10 +347,12 @@ class BatoceraManager(AbstractManager):
                 media_folder,
                 file_name
             )
-            FileHelper.copy_file(
-                source_file_path=media_file,
-                destination_file_path=batocera_media_files[media]
-            )
+
+            if Component.MEDIA in Context.get_selected_components():
+                FileHelper.copy_file(
+                    source_file_path=media_file,
+                    destination_file_path=batocera_media_files[media]
+                )
 
             # Add field for the media
             fields_to_add[key] = self.__FILE_PREFIX
@@ -353,7 +366,8 @@ class BatoceraManager(AbstractManager):
             if self.__FILE_PREFIX in line or line.strip() == "":
                 continue
             stripped = line.lstrip()
-            if stripped.startswith(f'<{self.__TAG_GAME}') or stripped.startswith(f'</{self.__TAG_GAME}'):
+            if stripped.startswith(f'<{self.__TAG_GAME}') or \
+                    stripped.startswith(f'</{self.__TAG_GAME}'):
                 lines.append(self.__PARENT_PREFIX + stripped)
             else:
                 lines.append(self.__CHILD_PREFIX + stripped)
@@ -397,7 +411,9 @@ class BatoceraManager(AbstractManager):
             1
         )
 
-        return FileHelper.write_file(
+        FileHelper.write_file(
             file_path=game_list_xml_path,
             content=game_list_xml_content
         )
+
+        return True
